extract pose detection from EditorCanvas to PoseEditor.
